scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nats-py>=2.14.0",
    "psycopg2-binary>=2.9.11",
]
